Moved pixel.py into observers and fixed import.
