+++ raysect/optical/observer/pixel.py
